server/requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
click==8.0.1
clikit==0.6.2
cmdstanpy==0.9.5
convertdate==2.3.2
crashtest==0.3.1
cycler==0.10.0
Cython==0.29.24
ephem==4.0.0.2
fbprophet==0.7.1
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
gunicorn==20.1.0
hijri-converter==2.1.3
holidays==0.11.2
httpstan==4.5.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
korean-lunar-calendar==0.2.1
LunarCalendar==0.0.9
lz4==3.1.3
MarkupSafe==2.0.1
marshmallow==3.13.0
matplotlib==3.4.2
multidict==5.1.0
numpy==1.21.1
pandas==1.3.1
pastel==0.2.1
Pillow==8.3.1
psycopg2==2.9.1
pylev==1.4.0
PyMeeus==0.5.11
pyparsing==2.4.7
pysimdjson==3.2.0
pystan==2.19.1.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
setuptools-git==1.2
six==1.16.0
SQLAlchemy==1.4.22
tqdm==4.62.0
typing-extensions==3.10.0.0
webargs==7.0.1
Werkzeug==2.0.1
yarl==1.6.3
